+++ accounting/src/apideck_accounting_unify/models/expenselineitem.py
@@ -42,6 +42,8 @@ class ExpenseLineItemTypedDict(TypedDict):
     r"""The expense line item description"""
     billable: NotRequired[bool]
     r"""Boolean that indicates if the line item is billable or not."""
+    line_number: NotRequired[Nullable[int]]
+    r"""Line number of the resource"""
 
 
 class ExpenseLineItem(BaseModel):
@@ -79,6 +81,9 @@ class ExpenseLineItem(BaseModel):
     billable: Optional[bool] = None
     r"""Boolean that indicates if the line item is billable or not."""
 
+    line_number: OptionalNullable[int] = UNSET
+    r"""Line number of the resource"""
+
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
         optional_fields = [
@@ -92,6 +97,7 @@ class ExpenseLineItem(BaseModel):
             "tax_rate",
             "description",
             "billable",
+            "line_number",
         ]
         nullable_fields = [
             "tracking_categories",
@@ -100,6 +106,7 @@ class ExpenseLineItem(BaseModel):
             "subsidiary_id",
             "description",
             "total_amount",
+            "line_number",
         ]
         null_default_fields = []