+++ accounting/RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [python v0.14.0] accounting
 ### Releases
 - [PyPI v0.14.0] https://pypi.org/project/apideck-accounting-unify/0.14.0 - accounting
+
+## 2025-05-22 00:05:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.548.4 (2.605.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.14.1] accounting
+### Releases
+- [PyPI v0.14.1] https://pypi.org/project/apideck-accounting-unify/0.14.1 - accounting